auth_service/requirements.txt
fastapi
uvicorn
asyncpg
prometheus-client
pyjwt
passlib
bcrypt==4.0.1
python-multipart
